fix(controller): change from int to long

+++ src/MathUtil.java
@@ -1,14 +1,14 @@
 public class MathUtil {	
-	public static int add(int number1,int number2){
+	public static long add(int number1,int number2){
 		return number1+number2;
 	}
-	public static int sub(int number1, int number2){
+	public static long sub(int number1, int number2){
 		return number1-number2;
 	}
-	public static int mul(int number1, int number2){
+	public static long mul(int number1, int number2){
 		return number1*number2;
 	}
-	public static int div(int number1, int number2){
+	public static long div(int number1, int number2){
 		if (number2 == 0){
 			throw new IllegalArgumentException("Division by zero");
 		}
